Some fixes

+++ cocos2dx/platform/win8_metro/BasicLoader.cpp
@@ -252,7 +252,7 @@ void BasicLoader::LoadShader(
 		DX::ThrowIfFailed(
 			m_d3dDevice->CreateVertexShader(
 			bytecode->Data,
-			bytecode->Length,
+			uint32_t(bytecode->Length),
 			nullptr,
 			shader
 			)
